Move UI to correct folders
