+++ app/dashboard/lines/(admin)/requests/[id]/page.tsx
@@ -1,13 +1,13 @@
 "use server";
 
 import { notFound } from "next/navigation";
-import ReviewRequestLayout from "./ui/ReviewRequestLayout";
 import { getLineRequestById } from "../../../(moderator)/my-requests/[id]/data/actions";
 import { getAgencyById } from "./data/get-agency-by-id";
 import { Line } from "@/app/lib/definitions";
 import { addNotesToRequests } from "./data/add-notes-to-request";
 import { approveRequest } from "./data/approve-request";
 import { rejectRequest } from "./data/reject-request";
+import ReviewRequestLayout from "@/ui/dashboard/lines/requests/ReviewRequestLayout";
 
 export default async function Page(props: { params: Promise<{ id: string }> }) {
   const params = await props.params;
