+++ ui/dashboard/lines/history/HistoryRecordDetails.tsx
