+++ ui/dashboard/lines/history/historyrecordlayout.module.scss
